LayoutTests/js/dom/Promise-catch-expected.txt

Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS thisInInit === firstPromise is true
PASS firstPromise instanceof Promise is true
PASS secondPromise instanceof Promise is true
PASS firstPromise.catch(null) threw exception TypeError: Expected function or undefined as as first argument.
PASS firstPromise.catch(37) threw exception TypeError: Expected function or undefined as as first argument.
PASS thisInFulfillCallback === firstPromise is false
PASS thisInFulfillCallback === secondPromise is true
PASS result is "hello"
PASS result is "bye"
PASS fulfilled
PASS successfullyParsed is true

TEST COMPLETE
